コードレビュー

+++ app/controllers/items_controller.rb
@@ -23,10 +23,14 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+   if user_signed_in?
     item = Item.find(params[:id])
     item.destroy
     redirect_to root_path
+   else
+    redirect_to root_path
   end
+end
 
   private
